Renamed JobDetails page folder.
