+++ src/components/common/StyledButton/index.tsx
@@ -1,4 +1,4 @@
-import React, { FC, ReactElement } from 'react';
+import React, { FC } from 'react';
 import Button from '@material-ui/core/Button';
 import CreateIcon from '@material-ui/icons/Create';
 
@@ -6,7 +6,7 @@ type Props = {
   text: string;
 };
 
-const StyledButton: FC<Props> = ({ text }: Props): ReactElement => (
+const StyledButton: FC<Props> = ({ text }: Props) => (
   <Button variant="contained" color="primary" startIcon={<CreateIcon />}>{text}</Button>
 );
